feat: Add context menu to save selected text to A few words
This commit adds a new context menu item to the Plasmo extension called "Save to A few words". When the user selects text on a webpage and right-clicks, they will see this option in the context menu. Clicking on it will save the selected text to the extension's local storage.

The necessary permissions for the context menu functionality have also been added to the manifest file.